removing unused method and refactor in types
